auth_service/pyproject.toml
[project]
name = "TwitterClone-AuthService"
version = "0.1.0"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "bcrypt==4.3.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "fastapi==0.115.11",
    "fastapi-cli==0.0.7",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "psycopg2-binary>=2.9.10",
    "pyasn1==0.4.8",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings>=2.8.1",
    "pygments==2.19.1",
    "python-dotenv==1.0.1",
    "python-jose==3.4.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "sqlmodel==0.0.24",
    "starlette==0.46.1",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
]
